The footer is now just a plain fluid div container with a nav list.
It is no longer a fixed-bottom nav bar.

Now using CSS tricks it floats to the bottom of the page even when there
is no content on the main page.

When there is lots of content the footer will be off the bottom of the
brower window; hence it won't be eating up valuable screen space when not
needed.

This Stack Overflow answer was very handy:

https://stackoverflow.com/questions/10099422/flushing-footer-to-bottom-of-the-page-twitter-bootstrap/28713303#28713303