Fixed filtering albums spec JS issue.
The click_on directive (a JS action) was somehow buggering up the GET
parameters when submitting. I don't know why. But the easiest solution is to
have a new 'filter=true' paramater for AlbumsController#index which displays
the filter form, hence it removes the need for JS display "filters" in the spec
all-together.

All good now.