Track visibility for Albums with long track listings is fully done.
Albums with more than 20 tracks will only display the first twenty by default
with a fade effect over the bottom four tracks of the twenty. A "Show all"
button is provided to display all tracks at which a "Show less" button is then
provided to go back to the concise listing.

Tests and documentation all done.