invisible_captcha gem is now hooked into User sign-up form.
This gem helps prevent bot sign-up by setting up a honeypot field that will not
be displayed in browers but which is likely to be filled in by a bot. If a bot
fills in the honeypot field then user sign-up will not be processed, instead
the registration process will be aborted and the bot will be redirected to the
home page.