It is "NOT" possible to cache comment fragments.
Fragment caching is global; a fragment is the SAME for every user.

However comment rendering CAN differ per user.

If the current_user is the same as the comment user then a little trash
can symbol will be rendered for that comment. Fragment cacheing totally
breaks that trash-can rendering. Hence, caching needs to be removed for
comment fragments unfortunately. Performance will be a little worse, but
early testing indicates it is not that bad. Comments are limited to MAX
of 25 a time and grows via endless scrolling. That will limit the
performance loss.