Mailer now set reply-to field
Previously the application was setting the 'from:' field of emails from
the value of MAILER_SENDER, which was my personal Gmail address.
Nowadays MTAs are more strict in regards to treating emails with 'from:'
fields that say they from Gmail, but are sent from Mailgun, as SPAM.

Instead I now set the 'from:' field as coming from the postmaster of my
registered domain and I populate the 'reply-to:' field with my personal
Gmail address.

Hopefully this works in production.