Use ES6 class arrow functions to bind this
Certain jQuery callbacks were not triggering due to 'this' not being
bound to the class object. I expected the use of 'bind' to solve it, but
alas not (I don't know why). Instead of spending time going down that
rabbit hole it was simpler to just use class arrow functions, which are
the recommended approach anyway.

It works. Moving on.